requirements.txt
PyYAML==3.11
bottle==0.12.7
cffi==0.8.6
cryptography==0.6.1
pyOpenSSL==0.14
pycparser==2.10
six==1.8.0
